Adjusted tooltip timing, moved MutePermissionTooltips to all HelpBox components
